+++ packages/dash-table/src/core/components/IsolatedInput/index.tsx
@@ -1,4 +1,5 @@
 import React, { PureComponent } from 'react';
+import { KEY_CODES } from 'dash-table/utils/unicode';
 
 type Submit = (value: string | undefined) => void;
 
@@ -6,6 +7,7 @@ interface IProps {
     placeholder?: string;
     updateOnBlur?: boolean;
     updateOnSubmit?: boolean;
+    updateOnEnter: boolean;
     stopPropagation?: boolean;
     submit: Submit;
     value: string | undefined;
@@ -13,6 +15,7 @@ interface IProps {
 
 interface IDefaultProps {
     stopPropagation: boolean;
+    updateOnEnter: boolean;
     updateOnBlur: boolean;
     updateOnSubmit: boolean;
 }
@@ -26,6 +29,7 @@ type PropsWithDefaults = IProps & IDefaultProps;
 export default class IsolatedInput extends PureComponent<IProps, IState> {
     public static readonly defaultProps = {
         stopPropagation: false,
+        updateOnEnter: true,
         updateOnBlur: true,
         updateOnSubmit: true
     };
@@ -42,25 +46,39 @@ export default class IsolatedInput extends PureComponent<IProps, IState> {
         };
     }
 
-    handleChange = (ev: any) => this.setState({
-        value: ev.target.value
-    })
+    handleKeyDown = (e: React.KeyboardEvent) => {
+        const {
+            stopPropagation,
+           updateOnEnter 
+        } = this.propsWithDefaults;
+
+        if(stopPropagation) e.stopPropagation();
+        if(updateOnEnter&& e.keyCode === KEY_CODES.ENTER) {
+            this.submit();
+        } 
+    }
+
+    handleChange = (ev: any) => {
+        this.setState({
+            value: ev.target.value
+        })
+    }
 
     submit = () =>
         this.state.value !== this.props.value &&
         this.props.submit(this.state.value)
 
+
     render() {
         const {
             placeholder,
-            stopPropagation,
             updateOnBlur,
             updateOnSubmit
         } = this.propsWithDefaults;
 
         let props = {
             onBlur: updateOnBlur ? this.submit : undefined,
-            onKeyDown: stopPropagation ? (ev: any) => ev.stopPropagation() : undefined,
+            onKeyDown: this.handleKeyDown,
             onSubmit: updateOnSubmit ? this.submit : undefined
         };