+++ packages/dash-table/CHANGELOG.md
@@ -4,8 +4,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [3.1.0-rc20] - 2018-11-01
 ### Fixed
-- Fix performance degradation on load [#200](https://github.com/plotly/dash-table/pull/200) [#198](https://github.com/plotly/dash-table/issues/198)
-
+- Fix performance degradation on load [#208](https://github.com/plotly/dash-table/pull/208) [#200](https://github.com/plotly/dash-table/pull/200) [#198](https://github.com/plotly/dash-table/issues/198)
 
 ## [3.1.0-rc19] - 2018-11-01
 ### Changed